add a reusable tryParseJson function
